[AIRFLOW-6552] Move Azure classes to providers.microsoft package (#7158)
