+++ airflow/contrib/hooks/azure_container_volume_hook.py
@@ -16,39 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""This module is deprecated. Please use `airflow.providers.microsoft.azure.hooks.azure_container_volume`."""
 
-from azure.mgmt.containerinstance.models import AzureFileVolume, Volume
+import warnings
 
-from airflow.hooks.base_hook import BaseHook
+# pylint: disable=unused-import
+from airflow.providers.microsoft.azure.hooks.azure_container_volume import AzureContainerVolumeHook  # noqa
 
-
-class AzureContainerVolumeHook(BaseHook):
-    """
-    A hook which wraps an Azure Volume.
-
-    :param wasb_conn_id: connection id of a Azure storage account of
-        which file shares should be mounted
-    :type wasb_conn_id: str
-    """
-
-    def __init__(self, wasb_conn_id='wasb_default'):
-        self.conn_id = wasb_conn_id
-
-    def get_storagekey(self):
-        conn = self.get_connection(self.conn_id)
-        service_options = conn.extra_dejson
-
-        if 'connection_string' in service_options:
-            for keyvalue in service_options['connection_string'].split(";"):
-                key, value = keyvalue.split("=", 1)
-                if key == "AccountKey":
-                    return value
-        return conn.password
-
-    def get_file_volume(self, mount_name, share_name,
-                        storage_account_name, read_only=False):
-        return Volume(name=mount_name,
-                      azure_file=AzureFileVolume(share_name=share_name,
-                                                 storage_account_name=storage_account_name,
-                                                 read_only=read_only,
-                                                 storage_account_key=self.get_storagekey()))
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.microsoft.azure.hooks.azure_container_volume`.",
+    DeprecationWarning, stacklevel=2
+)
