+++ airflow/contrib/operators/azure_cosmos_operator.py
@@ -16,54 +16,14 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
+"""This module is deprecated. Please use `airflow.providers.microsoft.azure.operators.azure_cosmos`."""
 
-from airflow.contrib.hooks.azure_cosmos_hook import AzureCosmosDBHook
-from airflow.models import BaseOperator
-from airflow.utils.decorators import apply_defaults
+import warnings
 
+# pylint: disable=unused-import
+from airflow.providers.microsoft.azure.operators.azure_cosmos import AzureCosmosInsertDocumentOperator  # noqa
 
-class AzureCosmosInsertDocumentOperator(BaseOperator):
-    """
-    Inserts a new document into the specified Cosmos database and collection
-    It will create both the database and collection if they do not already exist
-
-    :param database_name: The name of the database. (templated)
-    :type database_name: str
-    :param collection_name: The name of the collection. (templated)
-    :type collection_name: str
-    :param document: The document to insert
-    :type document: dict
-    :param azure_cosmos_conn_id: reference to a CosmosDB connection.
-    :type azure_cosmos_conn_id: str
-    """
-    template_fields = ('database_name', 'collection_name')
-    ui_color = '#e4f0e8'
-
-    @apply_defaults
-    def __init__(self,
-                 database_name,
-                 collection_name,
-                 document,
-                 azure_cosmos_conn_id='azure_cosmos_default',
-                 *args,
-                 **kwargs):
-        super().__init__(*args, **kwargs)
-        self.database_name = database_name
-        self.collection_name = collection_name
-        self.document = document
-        self.azure_cosmos_conn_id = azure_cosmos_conn_id
-
-    def execute(self, context):
-        # Create the hook
-        hook = AzureCosmosDBHook(azure_cosmos_conn_id=self.azure_cosmos_conn_id)
-
-        # Create the DB if it doesn't already exist
-        if not hook.does_database_exist(self.database_name):
-            hook.create_database(self.database_name)
-
-        # Create the collection as well
-        if not hook.does_collection_exist(self.collection_name, self.database_name):
-            hook.create_collection(self.collection_name, self.database_name)
-
-        # finally insert the document
-        hook.upsert_document(self.document, self.database_name, self.collection_name)
+warnings.warn(
+    "This module is deprecated. Please use `airflow.providers.microsoft.azure.operators.azure_cosmos`.",
+    DeprecationWarning, stacklevel=2
+)
