+++ airflow/models/connection.py
@@ -281,16 +281,16 @@ class Connection(Base, LoggingMixin):
             from airflow.contrib.hooks.redis_hook import RedisHook
             return RedisHook(redis_conn_id=self.conn_id)
         elif self.conn_type == 'wasb':
-            from airflow.contrib.hooks.wasb_hook import WasbHook
+            from airflow.providers.microsoft.azure.hooks.wasb import WasbHook
             return WasbHook(wasb_conn_id=self.conn_id)
         elif self.conn_type == 'docker':
             from airflow.hooks.docker_hook import DockerHook
             return DockerHook(docker_conn_id=self.conn_id)
         elif self.conn_type == 'azure_data_lake':
-            from airflow.contrib.hooks.azure_data_lake_hook import AzureDataLakeHook
+            from airflow.providers.microsoft.azure.hooks.azure_data_lake import AzureDataLakeHook
             return AzureDataLakeHook(azure_data_lake_conn_id=self.conn_id)
         elif self.conn_type == 'azure_cosmos':
-            from airflow.contrib.hooks.azure_cosmos_hook import AzureCosmosDBHook
+            from airflow.providers.microsoft.azure.hooks.azure_cosmos import AzureCosmosDBHook
             return AzureCosmosDBHook(azure_cosmos_conn_id=self.conn_id)
         elif self.conn_type == 'cassandra':
             from airflow.providers.apache.cassandra.hooks.cassandra import CassandraHook
