+++ airflow/providers/microsoft/azure/operators/wasb_delete_blob.py
@@ -22,43 +22,49 @@ from airflow.providers.microsoft.azure.hooks.wasb import WasbHook
 from airflow.utils.decorators import apply_defaults
 
 
-class FileToWasbOperator(BaseOperator):
+class WasbDeleteBlobOperator(BaseOperator):
     """
-    Uploads a file to Azure Blob Storage.
+    Deletes blob(s) on Azure Blob Storage.
 
-    :param file_path: Path to the file to load. (templated)
-    :type file_path: str
     :param container_name: Name of the container. (templated)
     :type container_name: str
     :param blob_name: Name of the blob. (templated)
     :type blob_name: str
     :param wasb_conn_id: Reference to the wasb connection.
     :type wasb_conn_id: str
-    :param load_options: Optional keyword arguments that
-        `WasbHook.load_file()` takes.
-    :type load_options: dict
+    :param check_options: Optional keyword arguments that
+        `WasbHook.check_for_blob()` takes.
+    :param is_prefix: If blob_name is a prefix, delete all files matching prefix.
+    :type is_prefix: bool
+    :param ignore_if_missing: if True, then return success even if the
+        blob does not exist.
+    :type ignore_if_missing: bool
     """
-    template_fields = ('file_path', 'container_name', 'blob_name')
+
+    template_fields = ('container_name', 'blob_name')
 
     @apply_defaults
-    def __init__(self, file_path, container_name, blob_name,
-                 wasb_conn_id='wasb_default', load_options=None, *args,
+    def __init__(self, container_name, blob_name,
+                 wasb_conn_id='wasb_default', check_options=None,
+                 is_prefix=False, ignore_if_missing=False,
+                 *args,
                  **kwargs):
         super().__init__(*args, **kwargs)
-        if load_options is None:
-            load_options = {}
-        self.file_path = file_path
+        if check_options is None:
+            check_options = {}
+        self.wasb_conn_id = wasb_conn_id
         self.container_name = container_name
         self.blob_name = blob_name
-        self.wasb_conn_id = wasb_conn_id
-        self.load_options = load_options
+        self.check_options = check_options
+        self.is_prefix = is_prefix
+        self.ignore_if_missing = ignore_if_missing
 
     def execute(self, context):
-        """Upload a file to Azure Blob Storage."""
-        hook = WasbHook(wasb_conn_id=self.wasb_conn_id)
         self.log.info(
-            'Uploading %s to wasb://%s as %s',
-            self.file_path, self.container_name, self.blob_name,
+            'Deleting blob: %s\nin wasb://%s', self.blob_name, self.container_name
         )
-        hook.load_file(self.file_path, self.container_name,
-                       self.blob_name, **self.load_options)
+        hook = WasbHook(wasb_conn_id=self.wasb_conn_id)
+
+        hook.delete_file(self.container_name, self.blob_name,
+                         self.is_prefix, self.ignore_if_missing,
+                         **self.check_options)
