+++ scripts/ci/pylint_todo.txt
@@ -3,11 +3,11 @@
 ./airflow/contrib/auth/backends/google_auth.py
 ./airflow/contrib/auth/backends/kerberos_auth.py
 ./airflow/contrib/auth/backends/ldap_auth.py
-./airflow/contrib/hooks/azure_container_instance_hook.py
-./airflow/contrib/hooks/azure_container_volume_hook.py
-./airflow/contrib/hooks/azure_cosmos_hook.py
-./airflow/contrib/hooks/azure_data_lake_hook.py
-./airflow/contrib/hooks/azure_fileshare_hook.py
+./airflow/providers/microsoft/azure/hooks/azure_container_instance.py
+./airflow/providers/microsoft/azure/hooks/azure_container_volume.py
+./airflow/providers/microsoft/azure/hooks/azure_cosmos.py
+./airflow/providers/microsoft/azure/hooks/azure_data_lake.py
+./airflow/providers/microsoft/azure/hooks/azure_fileshare.py
 ./airflow/contrib/hooks/datadog_hook.py
 ./airflow/contrib/hooks/dingding_hook.py
 ./airflow/contrib/hooks/discord_webhook_hook.py
@@ -29,10 +29,10 @@
 ./airflow/providers/apache/spark/hooks/spark_submit.py
 ./airflow/providers/apache/sqoop/hooks/sqoop.py
 ./airflow/contrib/hooks/vertica_hook.py
-./airflow/contrib/hooks/wasb_hook.py
-./airflow/contrib/operators/adls_list_operator.py
-./airflow/contrib/operators/azure_container_instances_operator.py
-./airflow/contrib/operators/azure_cosmos_operator.py
+./airflow/providers/microsoft/azure/hooks/wasb.py
+./airflow/providers/microsoft/azure/operators/adls_list.py
+./airflow/providers/microsoft/azure/operators/azure_container_instances.py
+./airflow/providers/microsoft/azure/operators/azure_cosmos.py
 ./airflow/contrib/operators/dingding_operator.py
 ./airflow/contrib/operators/discord_webhook_operator.py
 ./airflow/providers/apache/druid/operators/druid.py
@@ -73,10 +73,10 @@
 ./airflow/contrib/operators/ssh_operator.py
 ./airflow/contrib/operators/vertica_operator.py
 ./airflow/contrib/operators/vertica_to_mysql.py
-./airflow/contrib/operators/wasb_delete_blob_operator.py
+./airflow/providers/microsoft/azure/operators/wasb_delete_blob.py
 ./airflow/contrib/operators/winrm_operator.py
 ./airflow/contrib/sensors/aws_glue_catalog_partition_sensor.py
-./airflow/contrib/sensors/azure_cosmos_sensor.py
+./airflow/providers/microsoft/azure/sensors/azure_cosmos.py
 ./airflow/contrib/sensors/bash_sensor.py
 ./airflow/contrib/sensors/celery_queue_sensor.py
 ./airflow/contrib/sensors/datadog_sensor.py
@@ -97,7 +97,7 @@
 ./airflow/contrib/sensors/sagemaker_training_sensor.py
 ./airflow/contrib/sensors/sagemaker_transform_sensor.py
 ./airflow/contrib/sensors/sagemaker_tuning_sensor.py
-./airflow/contrib/sensors/wasb_sensor.py
+./airflow/providers/microsoft/azure/sensors/wasb.py
 ./airflow/sensors/weekday_sensor.py
 ./airflow/hooks/dbapi_hook.py
 ./airflow/hooks/docker_hook.py
