+++ tests/providers/microsoft/azure/__init__.py
@@ -16,14 +16,3 @@
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-"""This module is deprecated. Please use `airflow.providers.microsoft.azure.sensors.azure_cosmos`."""
-
-import warnings
-
-# pylint: disable=unused-import
-from airflow.providers.microsoft.azure.sensors.azure_cosmos import AzureCosmosDocumentSensor  # noqa
-
-warnings.warn(
-    "This module is deprecated. Please use `airflow.providers.microsoft.azure.sensors.azure_cosmos`.",
-    DeprecationWarning, stacklevel=2
-)
