+++ tests/providers/microsoft/azure/operators/test_adls_list.py
@@ -21,7 +21,7 @@ import unittest
 
 import mock
 
-from airflow.contrib.operators.adls_list_operator import AzureDataLakeStorageListOperator
+from airflow.providers.microsoft.azure.operators.adls_list import AzureDataLakeStorageListOperator
 
 TASK_ID = 'test-adls-list-operator'
 TEST_PATH = 'test/*'
@@ -31,7 +31,7 @@ MOCK_FILES = ["test/TEST1.csv", "test/TEST2.csv", "test/path/TEST3.csv",
 
 class TestAzureDataLakeStorageListOperator(unittest.TestCase):
 
-    @mock.patch('airflow.contrib.operators.adls_list_operator.AzureDataLakeHook')
+    @mock.patch('airflow.providers.microsoft.azure.operators.adls_list.AzureDataLakeHook')
     def test_execute(self, mock_hook):
         mock_hook.return_value.list.return_value = MOCK_FILES
