+++ tests/providers/microsoft/azure/operators/test_azure_cosmos.py
@@ -25,8 +25,8 @@ import uuid
 
 import mock
 
-from airflow.contrib.operators.azure_cosmos_operator import AzureCosmosInsertDocumentOperator
 from airflow.models import Connection
+from airflow.providers.microsoft.azure.operators.azure_cosmos import AzureCosmosInsertDocumentOperator
 from airflow.utils import db
